@@ -3,16 +3,25 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@emotion/react": "^11.7.0",
+        "@emotion/styled": "^11.6.0",
         "@fullcalendar/core": "^5.10.1",
         "@fullcalendar/daygrid": "^5.10.1",
         "@fullcalendar/interaction": "^5.10.1",
         "@fullcalendar/react": "^5.10.1",
         "@material-ui/core": "^4.12.3",
+        "@mui/material": "^5.2.3",
+        "bootstrap": "^5.1.3",
         "chart.js": "^3.6.2",
         "date-fns": "^2.27.0",
+        "i": "^0.3.7",
+        "jquery": "^3.6.0",
         "mdbreact": "^5.2.0",
+        "npm": "^8.2.0",
+        "popper.js": "^1.16.1",
         "prop-types": "^15.7.2",
-        "react": "^16.13.1",
+        "react": "^16.14.0",
+        "react-bootstrap": "^2.0.3",
         "react-burger-menu": "^2.6.13",
         "react-chartjs-2": "^4.0.0",
         "react-collapsible-content": "^1.0.0",
