@@ -1,8 +1,19 @@
-import React, {useState} from 'react';
+import React, { useState } from 'react';
 import { Column } from 'simple-flexbox';
 import { createUseStyles, useTheme } from 'react-jss';
-import { Grid, Typography } from '@material-ui/core';
-import { Row } from 'simple-flexbox';
+import { Grid, Modal, Typography } from '@material-ui/core';
+import { FaMoneyCheckAlt } from 'react-icons/fa';
+import ModalCompo from './ModalCompo';
+// import { Button } from 'react-bootstrap';
+import './Expense.css';
+import { render } from 'preact';
+import Button from '@mui/material/Button';
+import TextField from '@mui/material/TextField';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
 
 const useStyles = createUseStyles((theme) => ({
     addButton: {
@@ -11,8 +22,54 @@ const useStyles = createUseStyles((theme) => ({
         fontSize: '20px !important',
         padding: '7px !important'
     },
+    modal: {
+        height: '50vh',
+        position: 'fixed',
+        right: 0,
+        left: 0,
+        bottom: 0,
+        top:0,
+        backgroundColor: "white",
+        border: `1px solid ${theme.color.lightGrayishBlue2}`,
+        borderRadius: 4,
+        margin: "60px 30px 30px 30px",
+    },
+    modalH2:{
+        color: '#000000',
+        fontSize: '16px',
+    },
+    category:{
+        fontSize: '16px',
+        color: '#000000',
+        fontWeight: 'bold'
+    },
+
+    title:{
+        fontSize: '16px',
+        color: '#000000',
+        fontWeight: 'bold'
+    },
+
+    tag:{
+        fontSize: '16px',
+        color: '#000000',
+        fontWeight: 'bold'
+    },
+
+    date:{
+        fontSize: '16px',
+        color: '#000000',
+        fontWeight: 'bold'
+    },
+
+    price:{
+        fontSize: '16px',
+        color: '#000000',
+        fontWeight: 'bold'
+    },
     container: {
-        backgroundColor: '#FFFFFF',
+        backgroundColor: 'rgb(108, 99, 255)',
+        color: "#FFFFFF",
         border: `1px solid ${theme.color.lightGrayishBlue2}`,
         borderRadius: 4,
         cursor: 'pointer',
@@ -47,68 +104,111 @@ const useStyles = createUseStyles((theme) => ({
 const TAGS = {
     PERSONAL: { text: 'PERSONAL', backgroundColor: '#FEC400', color: '#FFFFFF' },
     GROUP: { text: 'GROUP', backgroundColor: '#29CC97', color: '#FFFFFF' }
-};4
+};
 
-function AddExpenseComponent({ className = '', title, value }) {
+function AddExpenseComponent(props) {
     const theme = useTheme();
     const classes = useStyles({ theme });
-    const [items, setItems] = useState([
-        { title: 'Grocery', checked: false, tag: TAGS.PERSONAL, price: "5.99" },
-        {
-            title: 'Transport',
-            checked: false,
-            tag: TAGS.GROUP,
-            price: "8.99"
-        },
-        { title: 'Food', checked: true, tag: TAGS.PERSONAL, price: "10.99" }
-    ]);
-    const composedClassName = [classes.container, className].join(' ');
+    const composedClassName = [classes.container, props.className].join(' ');
+    const [modal, setModal] = useState(false);
+    const [modalShow, setModalShow] = React.useState(false);
 
     function onAddButtonClick() {
-        setItems((prev) => {
+        props.setItems((prev) => {
             const newItems = [...prev];
             newItems.push({
-                title: `Task ${newItems.length + 1}`,
-                checked: false,
-                tag: getNextTag()
+                icon: <FaMoneyCheckAlt />,
+                title: `New Task`,
+                tag: TAGS.PERSONAL,
+                date: 'Dec 12, 2021',
+                price: '5.99'
             });
             return newItems;
         });
     }
-    
 
-    function getNextTag(current = 'GROUP') {
-        const tagLabels = ['PERSONAL', 'GROUP' ];
-        const tagIndex = (tagLabels.indexOf(current) + 1) % 2;
-        return TAGS[tagLabels[tagIndex]];
+    function onClick() {
+        if (!modal) {
+            setModal(true);
+
+        }
+        else {
+            setModal(false)
+        }
+    }
+    function onClose() {
+        setModal(false);
     }
 
-    function renderAddButton() {
-        return (
-            <Row
-                horizontal='center'
-                vertical='center'
-                className={[classes.tagStyles, classes.addButton].join(' ')}
-                onClick={onAddButtonClick}
-            >
-                +
-            </Row>
-        );
-    }
+
+
+    // const ModalComponent = () => {
+    //     return (
+    //         <div className={classes.modal}>
+    //             <div className={classes.modalH2}>Add Expense</div>
+    //             <div className={classes.category}>Category</div>
+    //             <div className={classes.title}>Title</div>
+    //             <div className={classes.tag}>Tag</div>
+    //             <div className={classes.date}>Date</div>
+    //             <div className={classes.price}>Price</div>
+
+    //         </div>
+    //     )
+    // };
+
+    const [open, setOpen] = React.useState(false);
+
+    const handleClickOpen = () => {
+      setOpen(true);
+    };
+  
+    const handleClose = () => {
+      setOpen(false);
+    };
 
     return (
-        <Column flexGrow={1} className={composedClassName} horizontal='center' vertical='center'>
-            <Grid item xs container direction="column" className={classes.user}>
-                <Grid item className={classes.user}>
-                <Typography variant="h6" style={{fontSize: '30px'}}>
-                       $1800.50
+        <Column flexGrow={1} className={composedClassName} horizontal='center' vertical='center' onClick={onClick}>
+            {/* <Grid item xs container direction="column" className={classes.user}>
+                <Grid item className={classes.user} >
+                    <Typography variant="h6" style={{ fontSize: '30px' }}>
+                        +
                     </Typography>
-                    <Typography variant="body2">Total Expenses</Typography>
-                    
+                    <Typography variant="body2">Add New Expense</Typography>
+                    { modal ? <ModalComponent/> : null }
+
+
                 </Grid>
-            </Grid> 
+            </Grid> */}
+            <div>
+      <Button onClick={handleClickOpen} style={{border: "1px solid white", color: "white"}}>
+        Add New Expense
+      </Button>
+        <Dialog open={open} onClose={handleClose}>
+            <DialogTitle>Expenses</DialogTitle>
+            <DialogContent>
+            <DialogContentText>
+                To subscribe to this website, please enter your email address here. We
+                will send updates occasionally.
+            </DialogContentText>
+            <TextField
+                autoFocus
+                margin="dense"
+                id="name"
+                label="Email Address"
+                type="email"
+                fullWidth
+                variant="standard"
+            />
+            </DialogContent>
+            <DialogActions>
+            <Button onClick={handleClose}>Cancel</Button>
+            <Button onClick={handleClose}>Save</Button>
+            </DialogActions>
+        </Dialog>
+        </div>
         </Column>
         
+
     );
 }
 
