@@ -8,10 +8,21 @@ const CollaborateComponent = lazy(() => import('./collaborate'));
 const ExpenseComponent = lazy(() => import('./expense'));
 const ToDo=lazy(()=> import('./ToDo'));
 
-function PrivateRoutes() {
+const SignInComponent = lazy(() => import('./signin'));
+const SignUpComponent = lazy(() => import('./signup'));
+
+function AppRoutes() {
     return (
         <Suspense fallback={<LoadingComponent loading />}>
             <Switch>
+            {/* <div className="auth-wrapper">
+            <div className="auth-inner"> */}
+                <Route path={SLUGS.login} component={SignInComponent} />
+                <Route path={SLUGS.signup} component={SignUpComponent} />
+                <Route path={SLUGS.forgotPassword} render={() => <div>forgotPassword</div>} />
+                {/* </div>
+            </div> */}
+                {/* <Redirect to={SLUGS.login} /> */}
                 <Route exact path={SLUGS.dashboard} component={DashboardComponent} />
                 <Route exact path={SLUGS.todo} component={ToDo} />
                 <Route exact path={SLUGS.expenses} component={ExpenseComponent} />
@@ -20,8 +31,9 @@ function PrivateRoutes() {
                 
                 {/* <Redirect to={SLUGS.dashboard} /> */}
             </Switch>
+            
         </Suspense>
     );
 }
 
-export default PrivateRoutes;
+export default AppRoutes;
