@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import useWindowSize from 'hooks/useWindowSize';
 import PrivateSection from 'routes/PrivateSection';
-import PublicRoutes from 'routes/PublicRoutes';
+// import AppRoutes from 'routes/AppRoutes';
 
 function Routes() {
     const { pathname } = useLocation();
@@ -14,7 +14,8 @@ function Routes() {
     }, [pathname]);
 
     const isUserLoggedIn = true;
-    return isUserLoggedIn ? <PrivateSection /> : <PublicRoutes />;
+    // return isUserLoggedIn ? <PrivateSection /> : <PublicRoutes />;
+    return <PrivateSection />;
 }
 
 export default Routes;
